# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portals OpenApi Documentation",
        "description": "Portals Swagger OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Portals API Server"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login to get token",
                "description": "login to get token",
                "operationId": "login",
                "requestBody": {
                    "description": "Data for storing collection",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "default": "fake@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "default": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "success"
                                        },
                                        "token": {
                                            "type": "string",
                                            "default": "1|oBbCtlOy0bbZogdrBucE4oghfyv5v5mWkDWTIoB5d33aeca4"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "default": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "default": "Test User"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "default": "fake@mail.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "default": "2024-04-08T23:51:10.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "default": "2024-04-08T23:51:10.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with to get the authentication token",
                "name": "Token based",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}